ground/requirements.txt
django>=3.2<3.3
channels
channels_redis
selenium
djangorestframework
django-cors-headers
requests
pytz